+++ datacontract/export/exporter_factory.py
@@ -206,3 +206,7 @@ exporter_factory.register_lazy_exporter(
 exporter_factory.register_lazy_exporter(
     name=ExportFormat.iceberg, module_path="datacontract.export.iceberg_converter", class_name="IcebergExporter"
 )
+
+exporter_factory.register_lazy_exporter(
+    name=ExportFormat.custom, module_path="datacontract.export.custom_converter", class_name="CustomExporter"
+)